+++ src/cli/deps.ts
@@ -1,11 +1,13 @@
 import { sendMessageDiscord } from "../discord/send.js";
 import { logWebSelfId, sendMessageWhatsApp } from "../providers/web/index.js";
+import { sendMessageSignal } from "../signal/send.js";
 import { sendMessageTelegram } from "../telegram/send.js";
 
 export type CliDeps = {
   sendMessageWhatsApp: typeof sendMessageWhatsApp;
   sendMessageTelegram: typeof sendMessageTelegram;
   sendMessageDiscord: typeof sendMessageDiscord;
+  sendMessageSignal: typeof sendMessageSignal;
 };
 
 export function createDefaultDeps(): CliDeps {
@@ -13,6 +15,7 @@ export function createDefaultDeps(): CliDeps {
     sendMessageWhatsApp,
     sendMessageTelegram,
     sendMessageDiscord,
+    sendMessageSignal,
   };
 }